refactor: 자주쓰는 함수 분리
