+++ commands/02-migrateHumancity.ts
@@ -5,14 +5,10 @@ import axios from 'axios'
 import async from 'async'
 import os from 'os'
 import inquirer from 'inquirer'
-import ffmpeg from 'fluent-ffmpeg'
 import jwt from 'jsonwebtoken'
-import dayjs from 'dayjs'
-import utc from 'dayjs/plugin/utc.js'
-import advancedFormat from 'dayjs/plugin/advancedFormat.js'
 
-dayjs.extend(utc)
-dayjs.extend(advancedFormat)
+import { humangateFilenameParser } from '../lib/dates'
+import { getVideoLength, createVideoThumbnail } from '../lib/videos'
 
 export default inquirer
   .prompt([
@@ -63,7 +59,7 @@ export default inquirer
 
       async.eachSeries(files, async (file) => {
         const videoLength = await getVideoLength(path.join(answers.recordingFilePath, file))
-        const recordedAt = fileNameDateParser(file)
+        const recordedAt = humangateFilenameParser(file)
         await createVideoThumbnail(answers.recordingFilePath, file)
 
         const result = await axios({
@@ -93,35 +89,3 @@ export default inquirer
       console.log(error)
     }
   })
-
-function getVideoLength (filePath: string): Promise<number> {
-  return new Promise((resolve, reject) => {
-    ffmpeg(filePath)
-      .ffprobe((err, metadata) => {
-        if (err) reject(err)
-        resolve(Math.round(metadata.format.duration || 0))
-      })
-  })
-}
-
-function fileNameDateParser (filename: string, format: string = 'x') {
-  const datestring = filename
-    .replace('y', '-').replace('m', '-').replace('d(', ' ')
-    .replace('h', ':').replace('m', ':').replace('s).mp4', '')
-
-  return dayjs(datestring).utc().format(format)
-}
-
-function createVideoThumbnail (baseDir:string, fileName: string): Promise<void> {
-  return new Promise((resolve, reject) => {
-    ffmpeg(path.join(baseDir, fileName))
-      .screenshot({
-        count: 1,
-        filename: fileName.replace('mp4', 'jpg'),
-        folder: baseDir
-      })
-      .on('end', () => {
-        return resolve()
-      })
-  })
-}
